# [CL-2908] Improve back-web-api-docs CI job - Part 2