--- a/back/engines/commercial/content_builder/spec/acceptance/layouts_spec.rb
+++ b/back/engines/commercial/content_builder/spec/acceptance/layouts_spec.rb
@@ -25,7 +25,7 @@ resource 'ContentBuilderLayouts' do
           {
             data: {
               id: layout.id,
-              type: 'content_builder_layout',
+              type: 'content_builder_layout CAUSE TEST FAILURE',
               attributes: hash_including(
                 code: code,
                 created_at: match(time_regex),