--- a/back/lib/tasks/rspec_api_docs.rake
+++ b/back/lib/tasks/rspec_api_docs.rake
@@ -7,9 +7,9 @@ RSpec::Core::RakeTask.new('web_api:docs:generate' => :environment) do |t, _task_
   ENV['DOC_FORMAT'] = 'html'
   ENV['API_NAME'] = 'CitizenLab Front Web API'
   t.pattern = if CitizenLab.ee?
-    '{spec,engines/**/spec}/acceptance/**/*_spec.rb'
+    '{spec,engines/**/spec}/acceptance/**/categories_spec.rb'
   else
     '{spec,engines/free/*/spec}/acceptance/**/*_spec.rb'
   end
-  t.rspec_opts = ['--format RspecApiDocumentation::ApiFormatter --exclude-pattern="engines/{admin_api,public_api}/**/*_spec.rb" -t ~admin_api']
+  t.rspec_opts = ['--format RspecApiDocumentation::ApiFormatter --exclude-pattern="engines/{admin_api,public_api}/**/*_spec.rb" -t ~admin_api -t ~@document:false']
 end