## The good news

Passing `'--format RspecApiDocumentation::ApiFormatter ... ~document:false'` helps a lot! ([ref](https://github.com/zipmark/rspec_api_documentation/issues/100#issuecomment-546848649)) Now we are only running tests that produce documentation!

[Before](https://app.circleci.com/pipelines/github/CitizenLabDotCo/citizenlab/93614/workflows/b063817f-d124-4e7c-aed1-fdbd133748ee/jobs/218599):
```Ruby
bundle exec rake web_api:docs:generate
...
Finished in 6 minutes 26 seconds (files took 10.96 seconds to load)
1502 examples, 0 failures, 8 pending
```
[After](https://app.circleci.com/pipelines/github/CitizenLabDotCo/citizenlab/93628/workflows/699e2a47-9285-40b6-9bce-04ad7bef4065/jobs/218665):
```Ruby
bundle exec rake web_api:docs:generate
...
Finished in 4 minutes 34.4 seconds (files took 11.49 seconds to load)
1202 examples, 0 failures, 6 pending
```
= approx. 2 minutes faster >= 20% speed improvement :-)

We also change to running `back-web-api-docs` concurrently with other jobs, not requiring a `back-test` success beforehand, and now the entire `back` workflow [completes in 9 mins](https://app.circleci.com/pipelines/github/CitizenLabDotCo/citizenlab/93630/workflows/a3c5f040-6b4d-48a6-a477-baeb1d61ff89), which is about 50% of the current duration. 😎 

## The bad news

Resource class `medium` had no significant advantage over `small` (probably because 2nd CPU is not utilised).

Splitting off `public_api` and `admin_api` docs into separate CI job only improved the main `web-api-docs` [by about 20 to 30 seconds](https://app.circleci.com/pipelines/github/CitizenLabDotCo/citizenlab/93618/workflows/8f47df28-cd6c-4530-9252-48a2888271c5) = not worth the extra cost.

## In other news

Also see small changes in [CL-2908] Improve back-web-api-docs CI job - Part 1 #4129


[CL-2908]: https://citizenlab.atlassian.net/browse/CL-2908?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ